Fixed more compiler warnings.

+++ library/LocalRegions/QuadExp.cpp
@@ -2244,10 +2244,10 @@ namespace Nektar
                      "Geometric information is not set up");
 
             // set up block matrix system
-            int nbdry = NumBndryCoeffs();
-            int nint = m_ncoeffs - nbdry;
+            unsigned int nbdry = NumBndryCoeffs();
+            unsigned int nint = (unsigned int)(m_ncoeffs - nbdry);
             unsigned int exp_size[] = {nbdry,nint};
-            int nblks = 2;
+            unsigned int nblks = 2;
             returnval = MemoryManager<DNekScalBlkMat>::
                 AllocateSharedPtr(nblks,nblks,exp_size,exp_size);
                 //Really need a constructor which takes Arrays
